+++ packages/cra-template/template/src/App.js
@@ -8,7 +8,6 @@ import {
   ShellBar,
   ThemeProvider
 } from '@ui5/webcomponents-react';
-import React from 'react';
 import './App.css';
 
 function App() {