+++ packages/main/src/components/VariantManagement/index.tsx
@@ -205,11 +205,21 @@ const styles = {
     fontWeight: 'normal'
   },
   footer: {
-    margin: '0.4375rem 1rem 0.4325rem auto'
+    '& > :last-child': {
+      marginRight: 0
+    }
   },
   inputIcon: { cursor: 'pointer', color: ThemingParameters.sapContent_IconColor },
-  searchInput: { padding: '0.25rem 0.5rem 0.25rem 0.25rem' },
-  popover: { minWidth: '25rem' }
+  searchInput: { padding: '0.25rem 1rem' },
+  popover: {
+    minWidth: '25rem',
+    '&::part(content)': {
+      padding: 0
+    },
+    '&::part(header), &::part(footer)': {
+      padding: '0 1rem'
+    }
+  }
 };
 
 const useStyles = createUseStyles(styles, { name: 'VariantManagement' });
@@ -310,9 +320,6 @@ const VariantManagement = forwardRef((props: VariantManagementPropTypes, ref: Re
   const handleSaveManageViews = (e, payload) => {
     const { defaultView, updatedRows, deletedRows } = payload;
     let callbackProperties = { deletedVariants: [], prevVariants: [], updatedVariants: [], variants: [] };
-    if (typeof onSaveManageViews === 'function') {
-      onSaveManageViews(enrichEventWithDetails(e, callbackProperties));
-    }
     setSafeChildren((prev) =>
       prev
         .map((child: ComponentElement<any, any>) => {
@@ -347,6 +354,9 @@ const VariantManagement = forwardRef((props: VariantManagementPropTypes, ref: Re
         })
         .filter(Boolean)
     );
+    if (typeof onSaveManageViews === 'function') {
+      onSaveManageViews(enrichEventWithDetails(e, callbackProperties));
+    }
     handleManageClose();
   };
 
@@ -471,6 +481,7 @@ const VariantManagement = forwardRef((props: VariantManagementPropTypes, ref: Re
             footer={
               (showSaveBtn || !hideSaveAs || !hideManageVariants || showCancelButton) && (
                 <Bar
+                  className={classes.footer}
                   endContent={
                     <>
                       {!inErrorState && showSaveBtn && (
@@ -571,6 +582,7 @@ const VariantManagement = forwardRef((props: VariantManagementPropTypes, ref: Re
         )}
         {saveAsDialogOpen && (
           <SaveViewDialog
+            portalContainer={portalContainer}
             showShare={!hideShare}
             showApplyAutomatically={!hideApplyAutomatically}
             showSetAsDefault={!hideSetAsDefault}