+++ packages/main/src/internal/SplitterLayoutContext.ts
@@ -2,8 +2,10 @@ import { createContext } from 'react';
 
 interface ISplitterLayoutContext {
   vertical: boolean;
+  reset: boolean;
 }
 
 export const SplitterLayoutContext = createContext<ISplitterLayoutContext>({
-  vertical: true
+  vertical: true,
+  reset: false
 });